Separate active transaction management from AbstractDistributedTransactionManager and AbstractTwoPhaseCommitTransactionManager
